requirements.txt
python-dotenv
pandas
pandas-gbq==0.18.1
prefect==2.7.7
prefect-sqlalchemy==0.2.2
prefect-gcp[cloud_storage]==0.2.3
protobuf==4.21.11